E-Commerce_Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
alembic
python-dotenv
email-validator
pydantic-settings